+++ app/routes/webhooks.app.uninstalled.tsx
@@ -6,6 +6,8 @@ export const action = async ({ request }: ActionFunctionArgs) => {
 
   console.log(`Received ${topic} webhook for ${shop}`);
 
+  // Webhook requests can trigger multiple times and after an app has already been uninstalled.
+  // If this webhook already ran, the session may have been deleted previously.
   if (session) {
     await sessionStorage.deleteSession(session.id);
   }